Move all mock objects to testing/mocked/
