+++ tests/store/dim/ucx_test.py
@@ -10,7 +10,7 @@ from proxystore.serialize import deserialize
 from proxystore.serialize import serialize
 from proxystore.store.dim.ucx import UCXServer
 from proxystore.store.dim.ucx import UCXStore
-from testing.mocker_modules.ucx_mocker import MockEndpoint
+from testing.mocked.ucx import MockEndpoint
 from testing.utils import open_port
 
 ENCODING = 'UTF-8'
